MYR-721: modularized codes for the styles and interfaces of SimpleCard
