+++ src/components-v2/atoms/simple-card/SimpleCard.tsx
@@ -6,58 +6,12 @@ import CardContent from '@material-ui/core/CardContent';
 import CardMedia from '@material-ui/core/CardMedia';
 import IconButton from '@material-ui/core/IconButton';
 import Typography from '@material-ui/core/Typography';
-import {createStyles, makeStyles, useTheme} from '@material-ui/core/styles';
+import {useTheme} from '@material-ui/core/styles';
 import MoreVertIcon from '@material-ui/icons/MoreVert';
 
+import {useStyles, SimpleCardProps} from './';
+
 import classNames from 'classnames';
-
-const useStyles = makeStyles(() =>
-  createStyles({
-    root: {
-      display: 'flex',
-      width: 'fit-content',
-
-      /* Drop shadow */
-      boxShadow: `0px 2px 10px rgba(0, 0, 0, 0.05)`,
-      borderRadius: 10,
-    },
-    activated: {
-      border: `1px solid #6E3FC3`,
-      boxSizing: `border-box`,
-    },
-    details: {
-      display: 'flex',
-      flexDirection: 'row',
-    },
-    content: {
-      flex: '1 0 auto',
-      alignSelf: 'center',
-      paddingLeft: 0,
-    },
-    cover: {
-      width: 68,
-      height: 68,
-      margin: 20,
-      opacity: 0.9,
-      borderRadius: 5,
-    },
-  }),
-);
-
-interface SimpleCardProps {
-  activated: boolean;
-  /*
-   * imgUrl can be a local path or an external url
-   */
-  imgUrl: string;
-
-  /*
-   * check if user owned this experience card
-   */
-  isOwner: boolean;
-
-  onClick: () => void;
-}
 
 const SimpleCard = ({activated, imgUrl, isOwner, ...props}: SimpleCardProps): JSX.Element => {
   const classes = useStyles();
