+++ src/components-v2/atoms/simple-card/index.ts
@@ -1,3 +1,6 @@
 import {default as SimpleCard} from './SimpleCard';
+import {default as useStyles} from './SimpleCard.style';
+import {SimpleCardProps} from './simple-card.interfaces';
 
-export {SimpleCard};
+export {SimpleCard, useStyles};
+export type {SimpleCardProps};
